Move RenderParams.php to Http\Controllers\ namespace
